requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
gunicorn==20.0.4
psycopg2
django-environ==0.9.0